Fixes per review
