+++ apis/mai400-api/src/tests/tx.rs
@@ -17,29 +17,27 @@
 use mai400::*;
 use super::*;
 
-//TODO: Get/Verify actual checksums
-
 #[test]
 fn reset_good() {
     let mock = mock_new!();
 
     // Packet for RequestReset
     mock.write.return_value_for(
-        (vec![
+        vec![
             0x90, 0xEB, 0x5A, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0,
             0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0,
             0x0, 0x0, 0x0, 0x0, 0xD5, 0x1,
-        ]),
+        ],
         Ok(()),
     );
 
     // Packet for ConfirmReset
     mock.write.return_value_for(
-        (vec![
+        vec![
             0x90, 0xEB, 0xF1, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0,
             0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0,
             0x0, 0x0, 0x0, 0x0, 0x6C, 0x2,
-        ]),
+        ],
         Ok(()),
     );
 
@@ -58,7 +56,7 @@ fn reset_bad_request() {
 
     // Packet for ConfirmReset
     mock.write
-        .return_value_for((vec![0x90, 0xEB, 0, 0, 0xF1, 0, 0xE0, 0x2E]), Ok(()));
+        .return_value_for(vec![0x90, 0xEB, 0, 0, 0xF1, 0, 0xE0, 0x2E], Ok(()));
 
     let mai = MAI400 {
         conn: Connection {
@@ -75,7 +73,7 @@ fn reset_bad_confirm() {
 
     // Packet for RequestReset
     mock.write
-        .return_value_for((vec![0x90, 0xEB, 0, 0, 0x5A, 0, 0x64, 0xEF]), Ok(()));
+        .return_value_for(vec![0x90, 0xEB, 0, 0, 0x5A, 0, 0x64, 0xEF], Ok(()));
 
     let mai = MAI400 {
         conn: Connection {
@@ -91,11 +89,11 @@ fn set_mode_good() {
     let mock = mock_new!();
 
     mock.write.return_value_for(
-        (vec![
+        vec![
             0x90, 0xEB, 0x0, 0x1, 0x2, 0x0, 0x3, 0x0, 0x4, 0x0, 0x5, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0,
             0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0,
             0x0, 0x0, 0x0, 0x0, 0x8A, 0x1,
-        ]),
+        ],
         Ok(()),
     );
 
@@ -129,11 +127,11 @@ fn set_mode_sun_good() {
     let mock = mock_new!();
 
     mock.write.return_value_for(
-        (vec![
+        vec![
             0x90, 0xEB, 0x0, 0x8, 0x1, 0x0, 0xCD, 0xCC, 0xC, 0x40, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0,
             0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0,
             0x0, 0x0, 0x0, 0x0, 0x0, 0x69, 0x3,
-        ]),
+        ],
         Ok(()),
     );
 
@@ -167,11 +165,11 @@ fn set_gps_time_good() {
     let mock = mock_new!();
 
     mock.write.return_value_for(
-        (vec![
+        vec![
             0x90, 0xEB, 0x44, 0x92, 0x3C, 0x74, 0x47, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0,
             0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0,
             0x0, 0x0, 0x0, 0x0, 0x0, 0x48, 0x3,
-        ]),
+        ],
         Ok(()),
     );
 
@@ -202,11 +200,11 @@ fn set_rv_good() {
     let mock = mock_new!();
 
     mock.write.return_value_for(
-        (vec![
+        vec![
             0x90, 0xEB, 0x41, 0xCD, 0xCC, 0x8C, 0x3F, 0xCD, 0xCC, 0xC, 0x40, 0x33, 0x33, 0x53,
             0x40, 0xCD, 0xCC, 0x8C, 0x40, 0x0, 0x0, 0xB0, 0x40, 0x33, 0x33, 0xD3, 0x40, 0x92, 0x3C,
             0x74, 0x47, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x55, 0xD,
-        ]),
+        ],
         Ok(()),
     );
 
@@ -245,7 +243,7 @@ fn passthrough_good() {
 
     let msg: [u8; 40] = [0x00; 40];
 
-    mock.write.return_value_for((msg.to_vec()), Ok(()));
+    mock.write.return_value_for(msg.to_vec(), Ok(()));
 
     let mai = MAI400 {
         conn: Connection {
